Starting the partition
