+++ VS/EUtilities/EUtilities.ixx
@@ -5,6 +5,8 @@ module;
 
 export module EUtilities;
 
+export import :Windows;
+
 export namespace EUtilities
 {
 	void flushTampon();
@@ -22,14 +24,14 @@ export namespace EUtilities
 	// KEYBOARD AND MOUSE INPUT //
 	//////////////////////////////
 
-	void pressKey(const char key);
-	void pressKey(const int keyCode);
+	void pressKey(char key);
+	void pressKey(int keyCode);
 	void pressKey(INPUT& input);
-	void releaseKey(const char key);
-	void releaseKey(const int keyCode);
+	void releaseKey(char key);
+	void releaseKey(int keyCode);
 	void releaseKey(INPUT& input);
-	void fullKeyPress(const char key, unsigned int pressDuration = 20);
-	void fullKeyPress(const int keyCode, unsigned int pressDuration = 20);
+	void fullKeyPress(char key, unsigned int pressDuration = 20);
+	void fullKeyPress(int keyCode, unsigned int pressDuration = 20);
 	void fullKeyPress(INPUT& input, unsigned int pressDuration = 20);
 
 	void ctrlV();
@@ -59,12 +61,6 @@ export namespace EUtilities
 
 	// Simulate a right click
 	void rightClick(const int& holdTime = 30);
-
-	// Wait until user presses and release the specified virutal key
-	void waitForFullKeyPress(short vKey);
-	void waitForKeyPress(short vKey);
-	void waitForKeyRelease(short vKey);
-
 
 	// Create folder if it doesn't exist. Returns false if folder did not exist
 	bool ensureFolderExists(const std::string_view dirName);
